Fix consume job

+++ core/core/src/org/smallbox/faraway/core/game/module/world/WorldModuleSerializer.java
@@ -229,7 +229,7 @@ public class WorldModuleSerializer implements SerializerInterface {
                     break;
 
                 case "maturity":
-                    maturity = vn.parseInt(vn.getText());
+                    maturity = vn.parseDouble(vn.getText());
 //                    maturity = 1;
                     break;
             }
